Moved RebelArray code from OGFILE3.cpp to new file OGFILE_REBELS.cpp.
