+++ src/OGFILE_REBELS.cpp
@@ -1,0 +1,56 @@
+/*
+* Seven Kingdoms: Ancient Adversaries
+*
+* Copyright 1997,1998 Enlight Software Ltd.
+* Copyright 2017 Richard Dijk <microvirus.multiplying@gmail.com>
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 2 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*
+*/
+
+#include <OREBEL.h>
+#include <file_io_visitor.h>
+
+static char* create_rebel_func();
+
+
+//-------- Start of function RebelArray::write_file -------------//
+//
+int RebelArray::write_file(File* filePtr)
+{
+	return write_ptr_array(filePtr, sizeof(Rebel));
+}
+//--------- End of function RebelArray::write_file ---------------//
+
+
+//-------- Start of function RebelArray::read_file -------------//
+//
+int RebelArray::read_file(File* filePtr)
+{
+	return read_ptr_array(filePtr, sizeof(Rebel), create_rebel_func);
+}
+//--------- End of function RebelArray::read_file ---------------//
+
+
+//-------- Start of static function create_rebel_func ---------//
+//
+static char* create_rebel_func()
+{
+	Rebel *rebelPtr = new Rebel;
+
+	rebel_array.linkin(&rebelPtr);
+
+	return (char*) rebelPtr;
+}
+//--------- End of static function create_rebel_func ----------//
